Save Plugin object after updating plugin_name attr
Although the plugin_name attribute was being set on the Plugin instance
during a scrape operation, the Plugin instance was not being saved.
This meant that the change was did not persist in the database
representation.  This commit fixes the issue by saving to the database
after updating the plugin_name.